+++ src/main/java/warehouse/repository/CustomerRepository.java
@@ -10,14 +10,14 @@ public class CustomerRepository extends CrudRepository<Customer> {
     private static final String ID_PARAM = "customersId";
     private static final String DELETE_QUERY = "delete from Customer c where c.id = :" + ID_PARAM;
 
-    public void delete(Long id) {
+    public void delete(int id) {
         runInTransaction((session) ->
                 session.createQuery(DELETE_QUERY)
                         .setParameter(ID_PARAM, id)
                         .executeUpdate());
     }
 
-    public Customer findOne(Long id) {
+    public Customer findOne(Integer id) {
         return super.findOne(id, Customer.class);
     }
